+++ compiler-tests/src/test/data/diagnostic/fir/bindingcontainer/AbstractContainersCannotHaveProvides.kt
@@ -1,0 +1,12 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+
+@BindingContainer
+interface ConfusedContainer {
+  // This is not legal
+  @Provides fun <!BINDING_CONTAINER_ERROR!>provideInt<!>(): Int = 3
+
+  companion object {
+    // This is legal
+    @Provides fun provideBoolean(): Boolean = false
+  }
+}