+++ compiler-tests/src/test/data/diagnostic/fir/bindingcontainer/CannotAnnotateBothGraphAndContainer.kt
@@ -1,0 +1,7 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+
+<!BINDING_CONTAINER_ERROR!>@DependencyGraph<!>
+@BindingContainer
+interface ConfusedContainer {
+  @Binds val Int.bind: Number
+}